% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_reconcile.R
\name{arg_reconcile}
\alias{arg_reconcile}
\title{Reconcile an argument list to any function signature.}
\usage{
arg_reconcile(
  fn,
  ...,
  .args = alist(),
  .docall = FALSE,
  .wtlist = c(),
  .bklist = c(),
  .remap = list(),
  .warn = TRUE,
  .error = "default",
  .finish = identity
)
}
\arguments{
\item{fn}{A function to which an argument list needs to be adapted.}

\item{...}{An arbitrary list of named arguments (unnamed ones will be 
ignored)}

\item{.args}{A list or \code{alist} of named arguments, to be merged 
with \code{...}}

\item{.docall}{If set to \code{TRUE} will not only clean up the arguments but 
also execute \code{fn} with those arguments (\code{FALSE} by
default)}

\item{.wtlist}{Whitelist. If not empty, only arguments named here will be 
permitted, and only if they satisfy the conditions implied by
the other arguments. Evaluated before \code{remap}.}

\item{.bklist}{Blacklist. If not empty, arguments named here will be removed 
even if they satisfy the conditions implied by the other 
arguments. Evaluated before \code{.remap}.}

\item{.remap}{An optional named character vector or named list of character 
values for standardizing arguments that play the same role but
have different names in different functions. Evaluated after
\code{.bklist} and \code{.wtlist}.}

\item{.warn}{Whether to issue a warning message (default) when invalid 
arguments need to be discarded.}

\item{.error}{If specified, should be the object to return in the event of
error. This object will have the error as its 
\code{error} attribute. If not specified an ordinary error is
thrown with an added hint on the documentation to read for 
troubleshooting. Ignored if \code{.docall} is \code{FALSE}.}

\item{.finish}{A function to run on the result before returning it. Ignored 
if \code{.docall} is \code{FALSE}.}
}
\value{
Either a named list or the result of calling \code{fn} with the 
        supplied arguments
}
\description{
Adapt an argument list to a function excluding arguments that 
             will not be recognized by it, redundant arguments, and un-named
             arguments.
}
\examples{

arg_reconcile(p.adjust,bla='aaa',baz='xzzz',n=40,
                    q=c(.1,.2,.02,.3,1,0,1),.remap = c(q='p'))
arg_reconcile(import,file='hello.csv',file='goodbye.csv',
              bla='aaa',baz='xzzz',format='csv')

}
