% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_attrs.R
\name{cleanup_attrs}
\alias{cleanup_attrs}
\title{Cleanup a data.frame using attribute heuristics}
\usage{
cleanup_attrs(x, convert_factors = TRUE, convert_na = FALSE,
  gather_attrs = FALSE, exclude = "", to = c("character", "factor"),
  percent = 40, quiet = FALSE)
}
\arguments{
\item{x}{A data.frame.}

\item{convert_factors}{A length-one logical. Should columns with labels for every value be converted to factors? Default is TRUE.}

\item{convert_na}{A length-one logical. Should labeled values in colums with sparse labels be converted to NA? Default is FALSE.}

\item{gather_attrs}{A length-one logical. Should variable-level meta-data be move to data.frame attributes? Default is FALSE}

\item{exclude}{exclude A character vector of column names to excluded from conversion. Default is "".}

\item{to}{Either 'character' (default) or 'factor'; type to convert categorical columns to.}

\item{percent}{A length-one integer used as the cutoff when identifying sparsely labeled columns. Default is 30.}

\item{quiet}{A length-one logical. Should information about modified colums be printed? Default is FALSE.}
}
\value{
\code{x}, usually with some columns converted to factors or character vectors, and optionally with some values replaced with NA and/or variable-level attributes moved to data.frame attributes.
}
\description{
This function by default converts columns with labels for every value to factors and labeled values in sparsely labeled columns with NA; it optionally moves variable-level attributes to the data.frame level
}
\details{
\code{\link{import}} attempts to standardize the return value from the various import functions to the extent possible, thus providing a uniform data structure regardless of what import package or function is used. It achieves this by storing any optional variable-related attributes at the variable level. This has the advantage of not losing meta-data during import, but often results in categorical data being imported as labeled numbers rather than characters or factors.  This function can be used to guess which columns contain categorical data and convert them to factors or character vectors.
}
\examples{
# Since Stata doesn't handle factors well, categorical data is often stored
# as labeled numbers, as in this data set
e <- import("http://www.stata-press.com/data/r14/labelbook1.dta")
e <- e[, c(20:21, 936, 997)]
str(e)
summary(e)
# Use variable attributes to guess wich colums are categorical and convert them
g <- cleanup_attrs(e)
str(g)
summary(g)
# Sometimes the heuristic is too aggressive; colums can be explicitly excluded
h <- cleanup_attrs(e, exclude = "rk21", to = "factor")
str(h)
summary(h)
# Optionally detect likely NA values and/or move column attributes to data.frame attributes
i <- cleanup_attrs(e, convert_na = TRUE, gather_attrs = TRUE, to = "factor")
str(i)
summary(i)
}
\seealso{
\code{\link{gather_attrs}}
}

