% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_envelopes.R
\name{describe_envelopes}
\alias{describe_envelopes}
\title{Describe envelopes}
\usage{
describe_envelopes(...)
}
\arguments{
\item{...}{Optional. scalar or vector. Returns a dataframe filtered by the
io \code{envelopes} table columns specified as parameter}
}
\value{
\code{tibble} with rows associated with IO data object
}
\description{
The function downloads and returns a list or dataframe containing a
description for each requested species distribution envelope where a row
corresponds to a record. The function returns a dataframe for all envelopes
if no parameters are specified. The function filters returned envelopes
record by attributes corresponding to io table columns specified as
parameters (ie. \code{id}, \code{taxa_scientific_name}, \code{scenario}, \code{year}, etc)
with accepted values either being scalar or vector for single or multiple
records
}
\examples{
# Returns all available envelopes records in io
envelopes <- describe_envelopes()

# Returns all envelopes filtered by the column `original_source`
envelopes <- describe_envelopes(taxa_scientific_name = "Binto binto")

}
