% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_haven.R
\name{heuristic_factor}
\alias{heuristic_factor}
\title{Convert data.frame variables to factors using variable attributes}
\usage{
heuristic_factor(x, to = c("character", "factor"), exclude = "",
  quiet = FALSE)
}
\arguments{
\item{x}{A data.frame.}

\item{to}{Either 'factor' or 'character'.}

\item{exclude}{A character vector of column names to excluded from conversion.}
}
\value{
\code{x}, usually with some columns converted to factors or character vectors.
}
\description{
This function converts columns with labels for every value to factors
}
\details{
\code{\link{import}} attempts to standardize the return value from the various import functions to the extent possible, thus providing a uniform data structure regardless of what import package or function is used. It achieves this by storing any optional variable-related attributes at the variable level. This has the advantage of not losing meta-data during import, but often results in categorical data being imported as labeled numbers rather than characters or factors. This function can be used to guess which columns contain categorical data and convert them to factors or character vectors.
}
\examples{
e <- import("http://www.stata-press.com/data/r14/labelbook1.dta")
e <- e[, c(20:21, 936, 997)]
str(e)
g <- rio:::heuristic_factor(e)
str(g)
h <- rio:::heuristic_factor(e, to = "character", exclude = "rk21")
str(h)
}
\seealso{
\code{\link{cleanup_attrs}}, \code{\link{gather_attrs}}
}
\keyword{internal}

