% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_haven.R
\name{heuristic_na}
\alias{heuristic_na}
\title{Guess missing values in data.frame variables using variable attributes}
\usage{
heuristic_na(x, percent = 40, exclude = "", quiet = FALSE)
}
\arguments{
\item{x}{A data.frame.}

\item{exclude}{A character vector of column names to excluded from conversion.}
}
\value{
\code{x}, usually with some values replaced with NA.
}
\description{
This function guesses which values are missing and converts them to NA
}
\details{
\code{\link{import}} attempts to standardize the return value from the various import functions to the extent possible, thus providing a uniform data structure regardless of what import package or function is used. It achieves this by storing any optional variable-related attributes at the variable level. This has the advantage of not losing meta-data during import, but often requires post-processing to identify missing values and convert them to NA. This function attempts to automate this process by guessing which values are missing based on the column attributes. Specifically, labeled values in columns with lables attributes for which less than \code{percent} of the values are labeled are converted to NA.
}
\examples{
e <- import("http://www.stata-press.com/data/r14/labelbook1.dta")
e <- e[, c(20:21, 936, 997)]
str(e)
g <- rio:::heuristic_na(e)
str(g)
}
\seealso{
\code{\link{cleanup_attrs}}, \code{\link{gather_attrs}}
}
\keyword{internal}

