% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_directory.R
\name{import_directory}
\alias{import_directory}
\title{Import directory of files to a data frame}
\usage{
import_directory(path, file.type = NULL, ...)
}
\arguments{
\item{path}{A character string path name that contains a directory of files with identical structure to be imported.}

\item{\dots}{Additional arguments passed to \code{\link{import}}.}
}
\value{
A data frame (or a list of data frames if \code{rbind} is unsuccessful).
}
\description{
Use \code{\link{import_list}} to import a directory of files list of data frames from a vector of file names or from a multi-object file (Excel workbook, .Rdata file, zip directory, or HTML file)
}
\examples{
library('datasets')
export(mtcars[1:8,],   "mtcars1.csv")
export(mtcars[9:16,],  "mtcars2.csv")
export(mtcars[17:24,], "mtcars3.csv")
export(mtcars[25:32,], "mtcars4.csv")

str(import_directory(path = ".", file.type = ".csv"))

# cleanup
unlink("mtcars1.csv")
unlink("mtcars2.csv")
unlink("mtcars3.csv")
unlink("mtcars4.csv")

}
\seealso{
\code{\link{import_list}}
}
