% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isfiletext.R
\name{isfiletext}
\alias{isfiletext}
\title{Determine whether a file is "plain-text" or some sort of binary format}
\usage{
isfiletext(
  filename,
  maxsize = Inf,
  textbytes = as.raw(c(7:16, 18, 19, 32:255)),
  tf = TRUE
)
}
\arguments{
\item{filename}{Path to the file}

\item{maxsize}{Maximum number of bytes to read}

\item{textbytes}{Which characters are used by normal (though not necessarily 
just ASCII) text. To detect just ASCII, the following value
can be used: `as.raw(c(7:16,18,19,32:127))`}

\item{tf}{If `TRUE` (default) simply return `TRUE` when `filename` 
references a text-only file and `FALSE` otherwise. If set to
`FALSE` then returns the "non text" bytes found in the file.}
}
\value{
boolean or raw
}
\description{
Determine whether a file is "plain-text" or some sort of binary format
}
\examples{
library(datasets)
export(iris,"iris.yml")
isfiletext("iris.yml")
## TRUE

export(iris,"iris.sav")
isfiletext("iris.sav")
## FALSE
isfiletext("iris.sav", tf=FALSE)
## These are the characters found in "iris.sav" that are not printable text
## 02 00 05 03 06 04 01 14 15 11 17 16 1c 19 1b 1a 18 1e 1d 1f
}
