% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_envelope.R
\name{post_envelopes}
\alias{post_envelopes}
\title{Post envelopes}
\usage{
post_envelopes(
  geom_ewkt,
  category,
  taxa_scientific_name,
  year,
  scenario = NULL,
  threshold_desc = NULL
)
}
\arguments{
\item{geom_ewkt}{\code{string}. Polygon or Multipolygon geometry
as Extended Well Known Text (EWKT) format with WGS84 as coordinate system
(SRID = 4326)}

\item{category}{\code{string}. One of either values \code{current} or \code{projected}.
Describes if envelopes are modeled at \code{current} observations and environment
variables or \code{projected} in future conditions.}

\item{taxa_scientific_name}{\code{string}. Scientific name of organism for which
the distribution is modeled.}

\item{year}{\code{integer}. Reference year for \code{current} or \code{projected}
distribution envelope.}

\item{scenario}{\code{string} Optional. Name of the scenarion describing projected
environment variables.}

\item{threshold_desc}{\code{string} Optional. Name or short description for SDM
thresholding methodology to generate a binary.}
}
\value{
NULL when successful
}
\description{
The function upload a species distribution model's envelope with related
information described as arguments to be stored in IO database.
}
\examples{
# Insert a small enveloppe for the newly discovered species `Binto binto`
response <- post_envelopes(
 geom_ewkt = "SRID=4326;POLYGON((-76.355 37.009,-76.350 37.010,-76.349 37.001,-76.354 37.001,-76.355 37.009))",
 category = "current",
 taxa_scientific_name = "Binto binto",
 year = 2022,
 scenario = NULL,
 threshold_desc = NULL
)
}
