\docType{package}
\name{rio}
\alias{rio}
\alias{rio-package}
\title{A Swiss-army knife for data file I/O}
\description{
  The aim of rio is to make data file I/O in R as easy as
  possible. The swiss-army knife-style functions export()
  and import() provide painless data file I/O experience.
}
\details{
  Exporting data

  \code{\link{export}} function writes data frame or matrix
  into an external data file with the data file format
  determined from the file extension.

  Importing data

  \code{\link{import}} function reads data frame from an
  external data file with the data file format determined
  from the file extension.
}

