% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_reconcile.R
\name{arg_reconcile}
\alias{arg_reconcile}
\title{Reconcile an argument list to any function signature.}
\usage{
arg_reconcile(
  fn,
  ...,
  .args = alist(),
  .docall = FALSE,
  .remap = list(),
  .warn = TRUE
)
}
\arguments{
\item{fn}{A function to which an argument list needs to be adapted.}

\item{...}{An arbitrary list of named arguments (unnamed ones will be 
ignored)}

\item{.args}{A list or \code{alist} of named arguments, to be merged 
with \code{...}}

\item{.docall}{If set to \code{TRUE} will not only clean up the arguments but 
also execute \code{fn} with those arguments (\code{FALSE} by
default)}

\item{.remap}{An optional named character vector or named list of character 
values for standardizing arguments that play the same role but
have different names in different functions.}

\item{.warn}{Whether to issue a warning message (default) when invalid 
arguments need to be discarded.}
}
\value{
Either a named list or the result of calling \code{fn} with the 
        supplied arguments
}
\description{
Adapt an argument list to a function excluding arguments that 
             will not be recognized by it, redundant arguments, and un-named
             arguments.
}
\examples{

arg_reconcile(p.adjust,bla='aaa',baz='xzzz',n=40,
                    q=c(.1,.2,.02,.3,1,0,1),.remap = c(q='p'))
arg_reconcile(import,file='hello.csv',file='goodbye.csv',
              bla='aaa',baz='xzzz',format='csv')
              
}
