% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/export.R
\name{convert}
\alias{convert}
\title{Convert from one file format to another}
\usage{
convert(in_file, out_file, in_opts = list(), out_opts = list())
}
\arguments{
\item{in_file}{a character string naming an input file.}

\item{out_file}{a character string naming an output file.}

\item{in_opts}{a named list of options to be passed to \code{\link{import}}}

\item{out_opts}{a named list of options to be passed to \code{\link{export}}}
}
\value{
The name of the output file (invisibly).
}
\description{
This function constructs a data frame from a data and exports to the specified format based on the file extension.
}
\examples{
export(iris, "iris.csv")
convert("iris.csv", "iris.dta")
import("iris.dta")
}

