% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ext}
\alias{get_ext}
\title{Get File Type from Extension}
\usage{
get_ext(file)
}
\arguments{
\item{file}{A character string containing a filename, file path, or URL.}
}
\value{
A characters string containing a file type recognized by rio.
}
\description{
A utility function to retrieve the file type from a file extension (via its filename/path/URL)
}
\examples{
get_ext("starwars.xlsx")
get_ext("starwars.ods")
get_ext("clipboard") ## "clipboard"
get_ext("https://github.com/ropensci/readODS/raw/v2.1/starwars.ods")
}
