% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_list.R
\name{import_list}
\alias{import_list}
\title{Import list of data frames}
\usage{
import_list(file, which, ...)
}
\arguments{
\item{file}{A character string containing a single file name for a multi-object file (e.g., Excel workbook, zip directory, or HTML file), or a vector of file paths for multiple files to be imported.}

\item{which}{If \code{file} is a single file path, this specifies which objects should be extracted (passed to \code{\link{import}}'s \code{which} argument). Ignored otherwise.}

\item{\dots}{Additional arguments passed to \code{\link{import}}. Behavior may be unexpected if files are of different formats.}
}
\value{
A list of a data frames.
}
\description{
Use \code{\link{import}} to import a list of data frames from a vector of file names or from a multi-object file (Excel workbook, .Rdata file, zip directory, or HTML file)
}
\examples{
library('datasets')
export(list(iris = iris, mtcars = mtcars), "data.xlsx")

# import a single file from multi-object workbook
str(import("data.xlsx", which = "mtcars"))

# import all worksheets
str(import_list("data.xlsx"))

# import multiple files
export(mtcars, "mtcars.csv")
export(mtcars, "iris.csv")
str(import_list(c("mtcars.csv", "iris.csv")))

# cleanup
unlink("data.xlsx")
unlink("mtcars.csv")
unlink("iris.csv")

}
\seealso{
\code{\link{import}}
}
