% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isfiletext.R
\name{is_file_text}
\alias{is_file_text}
\title{Determine whether a file is "plain-text" or some sort of binary format}
\usage{
is_file_text(file, maxsize = Inf, text_bytes = as.raw(c(7:16, 18, 19, 32:255)))
}
\arguments{
\item{file}{Path to the file}

\item{maxsize}{Maximum number of bytes to read}

\item{text_bytes}{Which characters are used by normal text (though not 
necessarily just ASCII). To detect just ASCII, the 
following value can be used: 
\code{as.raw(c(7:16, 18, 19, 32:127))}}
}
\value{
boolean
}
\description{
For debugging (i.e. why is a file being deemed non-text) set the option
\code{"rio.is_file_text.debug"} to \code{TRUE}. Then the result returned will 
also have an attribute named \code{"non_text_bytes"} which will be a vector 
of the raw bytes which prevented the file from being identified as text
}
\examples{
library(datasets)
export(iris, "iris.yml")
is_file_text("iris.yml")
## TRUE

export(iris, "iris.sav")
is_file_text("iris.sav")
## FALSE

}
