\docType{package}
\name{rio}
\alias{rio}
\alias{rio-package}
\title{A Swiss-army knife for data file I/O}
\description{The aim of rio is to make data file I/O in R as easy as possible. The \code{export} and \code{import} serve as a Swiss-army knife for painless data I/O for data from almost any file format.}
